fixes #3021, refs #1331 - adds Piwik_View_OneClickDone
During a Piwik software update, there will be instances of old classes
loaded in memory.  This is problematic as we will start to instantiate
new classes which may not be backward compatible.  This class provides
a clean bridge/transition by forcing a new request.

This class needs to be self-contained, with no external dependencies.